% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_color_return_rgb}
\alias{check_color_return_rgb}
\title{Checks color variables are either RGB values, hex color codes or a recognised color name
and converts to rgb (helper funct)}
\usage{
check_color_return_rgb(color, color_variable_name)
}
\arguments{
\item{color}{The color string / rgb vector to check}

\item{color_variable_name}{The name of the variable, for readability of error messages}
}
\value{
An error message if the color value can't be interpreted
}
\description{
Checks color variables are either RGB values, hex color codes or a recognised color name
and converts to rgb (helper funct)
}
\examples{
check_color_return_rgb("White", "test_color")
\dontrun{check_color_return_rgb("foo", "test_color")}

}
