% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_colour_return_rgb}
\alias{check_colour_return_rgb}
\title{Checks colour variables are either RGB values, hex colour codes or a recognised colour name
and converts to rgb (helper funct)}
\usage{
check_colour_return_rgb(colour, colour_variable_name)
}
\arguments{
\item{colour}{The colour string / rgb vector to check}

\item{colour_variable_name}{The name of the variable, for readability of error messages}
}
\value{
An error message if the colour value can't be interpreted
}
\description{
Checks colour variables are either RGB values, hex colour codes or a recognised colour name
and converts to rgb (helper funct)
}
\examples{
check_colour_return_rgb("White", "test_colour")
\dontrun{check_colour_return_rgb("foo", "test_colour")}

}
