% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monochromeR-package.R
\docType{package}
\name{monochromeR-package}
\alias{monochromeR}
\alias{monochromeR-package}
\title{monochromeR: Create and View Monochrome Colour Palettes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Let's you create and view monochrome colour palettes. Two options are given to create palettes, go towards black and white from your start colour or "blend" your start colour with another colour.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cararthompson/monochromeR}
  \item Report bugs at \url{https://github.com/cararthompson/monochromeR/issues}
}

}
\author{
\strong{Maintainer}: Cara Thompson \email{packages@cararthompson.com} (\href{https://orcid.org/0000-0002-6472-5073}{ORCID})

}
\keyword{internal}
