% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-monochrome.R
\name{palette_monochrome_blend}
\alias{palette_monochrome_blend}
\title{Generate monochrome palette from start colour going towards another colour}
\usage{
palette_monochrome_blend(colour, blend_colour, n_colours)
}
\arguments{
\item{colour}{The colour starting from which the palette is to be generated.
Must be a length one character of a specification that \code{\link[prismatic:clr_mix]{prismatic::clr_mix()}}
accepts as \code{col}.}

\item{blend_colour}{The colour to be used for "blending". Must be a length
one character of a specification that \code{\link[prismatic:clr_mix]{prismatic::clr_mix()}} accepts as \code{col}.}

\item{n_colours}{A length 1 integerish (but not factor) vector indicating the
number of colours to generate for the palette.}
}
\value{
A length \code{n_colours} character vector with hexencoded rgb values
(i.e., "#rrggbb").
}
\description{
\code{palette_monochrome_blend()} generates a monochrome palette from a start
colour going towards another colour. It is a wrapper around
\code{\link[prismatic:clr_mix]{prismatic::clr_mix()}}.
}
\details{
\itemize{
\item Alpha values in \code{colour} and \code{blend_colour} are ignored.
\item The colour specification refers to the standard sRGB colorspace (IEC
standard 61966).
}
}
\examples{
palette_monochrome_blend("#0D3F53", blend_colour = "white", n_colours = 5)
palette_monochrome_blend("#96410EFF", blend_colour = "#418979FF", n_colours = 6)

}
\seealso{
Other monochrome palettes: 
\code{\link{palette_monochrome_bw}()}
}
\concept{monochrome palettes}
