% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-monochrome.R
\name{palette_monochrome_bw}
\alias{palette_monochrome_bw}
\title{Generate monochrome palette from start colour going towards black and white}
\usage{
palette_monochrome_bw(colour, n_colours)
}
\arguments{
\item{colour}{The colour starting from which the palette is to be generated.
Must be a length one character of a specification that \code{\link[prismatic:clr_mix]{prismatic::clr_mix()}}
accepts as \code{col}.}

\item{n_colours}{A length 1 integerish (but not factor) vector. Must be an
odd number, so that the created palette is balanced in the sense of
containing \code{floor(n_colours / 2)} colours in both directions. Must be larger
than 1.}
}
\value{
A character vector of length \code{n_colours} with hexencoded rgb values
(i.e., "#rrggbb").
}
\description{
\code{palette_monochrome_blend()} generates a monochrome palette from a start
colour going towards black and white.
}
\details{
\itemize{
\item Alpha values in \code{colour} are ignored.
\item The colour specification refers to the standard sRGB colorspace (IEC
standard 61966).
}
}
\examples{
palette_monochrome_bw("violetred3", n_colours = 5)
palette_monochrome_bw("#CD3278FF", n_colours = 5)
palette_monochrome_bw("#0F4B63", n_colours = 11)

}
\seealso{
Other monochrome palettes: 
\code{\link{palette_monochrome_blend}()}
}
\concept{monochrome palettes}
