% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rgb_to_hex}
\alias{rgb_to_hex}
\title{Converts RGB values to hex color code}
\usage{
rgb_to_hex(x)
}
\arguments{
\item{x}{A matrix of red, blue and green values}
}
\value{
A corresponding hex color code
}
\description{
Converts RGB values to hex color code
}
\examples{
temp_rgb_matrix <- rgba_to_rgb(c(52, 46, 39, 0.8))
rgb_to_hex(temp_rgb_matrix)

}
